use go/packages
This has a few advantages.

It gets us out of the business of parsing 'go list' output
and managing walking the dependency tree.

It allows us to accept multiple packages and/or package
patterns, like './...' (or 'std'!).
This was my primary motivation:
I wanted to be able to write 'depaware -check ./tailscale{,d}'.

It makes it easier to add support for test packages
and other more sophisticated analysis later.

Fixes #2